core/requirements.txt
langchain
google-generativeai
python-dotenv
langchain_google_genai
streamlit
chromadb